pyproject.toml
[project]
name = "own-chatgpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "async>=0.6.2",
    "chainlit>=2.5.5",
    "openai>=1.93.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
